Move withServices to context, remove React.createContext ponyfill in lieu of ServiceConsumer
